# docs/api_swagger.yaml
openapi: 3.0.3
info:
  title: Projeto monitoramento em transporte de pets - OpenAPI 3.0
  description: |-
      Esta é a documentação da API para comunicação com o backend do projeto
      
      **Esta API requer autenticação com JWT**
      

       
  version: v1
servers:
  - url: http://localhost/v1/
...
paths:
...
  /user/{userid}:
    get:
      tags:
        - user
      summary: Recupera os dados do usuário com o id indicado
      description: ''
      operationId: getUserByName
      parameters:
        - name: userid
          in: path
          description: 'Id do usuário desejado'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGET'          

        '404':
          description: Userario não encontrado
          
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []
        
    put:
      tags:
        - user
      summary: Atualizar usuário
      description: Atualiza os dados do usuário indicado
      operationId: updateUser
      parameters:
        - name: userid
          in: path
          description: Id do usuário a ser modificado
          required: true
          schema:
            type: string
      requestBody:
        description: Atualiza um usuário existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        default:
          description: Atualizado com sucesso
          
      security:
        - bearerAuth: []
          
    delete:
      tags:
        - user
      summary: Remove um usuário
      description: Remove o usuário com id indicado
      operationId: deleteUser
      parameters:
        - name: userid
          in: path
          description: Id do usuário
          required: true
          schema:
            type: string
      responses:
        '404':
          description: Usuário não encontrado
          
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []  
...
components:
  schemas:
...
    User:
      type: object
      properties:
        nome:
          type: string
          example: Juca
        documento:
          type: string
          example: '11.208.655.95'
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
    UserGET:
      type: object
      properties:
        id:
          type: string
          example: 90fce5c3-22a7-4b4f-822f-1bb34a36c75d
        nome:
          type: string
          example: Juca
        documento:
          type: string
          example: '11.208.655.95'
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
...
  responses:
    UnauthorizedError:
      description: Operação não autorizada, verifique suas permissões
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes